Progress

+++ Capstone/java/src/main/java/com/techelevator/Candy.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends InventoryItem{
 
-    public Candy( String name, double price) {
+    public Candy(String name, BigDecimal price) {
         super(name, price);
     }
 }
