+++ Capstone/java/src/main/java/com/techelevator/Gum.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Gum extends InventoryItem{
 
-    public Gum(String product, double price) {
+    public Gum(String product, BigDecimal price) {
         super(product, price);
     }
 }
